+++ spec/spec_helper.rb
@@ -13,6 +13,7 @@ require "linzer"
 require_relative "rfc9421_examples"
 require_relative "rack_helper"
 require_relative "request_helper"
+require_relative "faraday_helper"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure