+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Enforce expires signature parameter validation in Verifier.
 
+- Add Faraday middleware and adapters for HTTP message signatures.
+
 ## [0.7.9.beta2] - 2026-04-19
 
 - Add support for http gem 6.x while maintaining compatibility with 5.x.